dhewm 3 is a Doom 3 GPL source modification.
The goal of dhewm 3 is bring DOOM 3 with the help of SDL to all
suitable plaforms.

This source release does not contain any game data, the game data
is still covered by the original EULA and must be obeyed as usual.

You must patch the game to the latest version.

ok sthen@