Import libv4l 1.8.1
libv4l is a collection of libraries which adds a thin abstraction layer on top
of video4linux2 devices. The purpose of this (thin) layer is to make it easy
for application writers to support a wide variety of devices without having to
write separate code for different devices in the same class.

OK ajacoutot@, czarkoff@

Written by: feinerer <feinerer@cvs.openbsd.org>