Add an initial pledge to mutt, from tb@. Tested so far with slang/normal, gpgme/normal, sasl/normal, headercache on/off, and with imap/mbox/maildir.
As you'd expect with a program like this that has many configuration
options, including at runtime, the pledge is rather wide. But even so,
a wide pledge still prevents many system calls, and restricts use of
others.

Committing early to make it easier for people to test. Please do so,
especially if you have an unusual configuration. If you run into problems,
please obtain a ktrace and report back to tb@ and myself with a description,
the last page or so of kdump output, and preferably a backtrace. Thanks!