Import py-alabaster 0.7.7, ok sthen@
Alabaster is a visually (c)lean, responsive, configurable theme for the
Sphinx documentation system. It is Python 2+3 compatible.

It began as a third-party theme, and is still maintained separately, but
as of Sphinx 1.3, Alabaster is an install-time dependency of Sphinx and
is selected as the default theme.

Written by: Alexandr Shadchin <shadchin@openbsd.org>