Import lsyncd 2.1.6.
Lsyncd watches a local directory tree using libinotify.  It aggregates
and combines events for a few seconds and then spawns one (or more)
process(es) to synchronize the changes. By default this is rsync. Lsyncd
is thus a light-weight live mirror solution that is comparatively easy
to install not requiring new filesystems or blockdevices and does not
hamper local filesystem performance.

ok ajacoutot@