Import google-cloud-sdk-91.0.1.
The Google Cloud SDK contains tools and libraries that allow you to create and
manage resources on Google's Cloud Platform, including App Engine, Compute
Engine, Cloud Storage, Cloud SQL, and BigQuery.

Note that for obvious reasons, the "docker" subcommand and kubectl are not
supported on OpenBSD for managing a GKE cluster.

ok sthen@