security update to PuTTY 0.67
This release fixes a security hole in PSCP, in the old-style SCP
protocol. A server sending a malformed header before the contents of the
file could overrun a buffer exploitably in PSCP. [CVE-2016-2563]

plus "Assorted other fixes for crash-type bugs (but none known to be
exploitable)".

Written by: Stuart Henderson <sthen@openbsd.org>