import p5-NetAddr-MAC, from Mikolaj Kucharski, ok benoit@
NetAddr::MAC module provides an interface to deal with Media Access
Control addresses. These are the addresses that uniquely identify a
device on various layer 2 networks. Although the most common case is
hardware addresses on Ethernet network cards, there are a variety of
devices that use this system of addressing.

This module supports both EUI-48 and EUI-64 addresses and implements
an object-oriented and a functional interface.