Update to poppler-0.42.0.
Not including pdfsig(1) and related functionaliety in the
libraries for now, because it would pull in a couple of
additional dependencies (via security/nss). I'll send a
separate diff for this in a few days to let people judge
wether it's too much or not.

ok aja@

Written by: Matthias Kilian <kili@openbsd.org>