import p5-Test-Portability-Files 0.06 OK benoit@ okan@ sthen@
This module is used to check the portability across operating systems
of the names of the files present in the distribution of a module.
The tests use the advices given in "Files and Filesystems" in
perlport.  The author of a distribution can select which tests to
execute.

Written by: Alexander Bluhm <bluhm@openbsd.org>