Import py-shapely 1.5.15.
Shapely is a BSD-licensed Python package for manipulation and analysis of
planar geometric objects. It is based on the widely deployed GEOS (the engine
of PostGIS) and JTS (from which GEOS is ported) libraries. Shapely is not
concerned with data formats or coordinate systems, but can be readily
integrated with packages that are.

ok sthen@

Written by: Landry Breuil <landry@openbsd.org>