Stop using sigreturn from the Go openbsd/386 signal trampoline - the sigreturn syscall is going away in future versions of OpenBSD and its use for Go is already broken due to the trampoline return checks. Instead just return directly from the signal trampoline (as we have always done for Go openbsd/amd64).
Also fix a bug in the Go openbsd/386 linker, which results in symbols in
the dynamic symbol table being emitted with a size of zero. With a current
ld.so, this results in symbol mismatch warnings.

Unbreaks the Go port on openbsd/386, however it still fails to pass regress
due to the PT_TLS changes.

Written by: Joel Sing <jsing@openbsd.org>