Import KDE Frameworks 5.21.0. Somewhat outdated, but we need to move on.
This does not include NetworkManager and ModemManager bindings, for obvious
reasons. But there is the Bluez-Qt binding included, which acts as a stub,
allowing KDE stuff relying on it to compile.

Also, we ship outdated version of Baloo, because newer versions rely on
the MDB library which we don't have - and won't have until UBC is fixed.

Input from many, tested by kirby@ and awolk@, final okay awolk@ - thanks!

Written by: zhuk <zhuk@cvs.openbsd.org>