Import emulators/hatari from Frederic Cambus
Hatari is an Atari ST/STE/TT/Falcon emulator for Linux, FreeBSD, NetBSD,
BeOS, Mac-OSX and other Systems which are supported by the SDL library.
Unlike most other open source ST emulators which try to give you a good
environment for running GEM applications, Hatari tries to emulate the
hardware as close as possible so that it is able to run most of the old
Atari games and demos.  Because of this, it may be somewhat slower than
less accurate emulators.

OK bentley@ on initial tarball
OK awolk@, abieber@ on latest tarball


Written by: awolk <awolk@cvs.openbsd.org>